Add feature to change default currency
The settings for the default currency is sored in JSON format in a file
locally. The values is retireved at runtime from the file. And if option
selected, then the new default value will be stored in them. Removed the
requirements and test-requirements file, as this information is already
in the setup.py. Updated the README.rst with the usage of the new
commands, and fixed the typo.