Add offline support and show list of currencies
Add supporting functions to show list of currencies available and add
offline support by storing the exchange rates in a json file.

Remove the json files for testing, and use the original files instead.

Not commiting exhcange rates file, since we have a function to download
the rates if not found.

This closes #24, fixes #16.